requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
hikari==2.0.0.dev115
hikari-lightbulb==2.3.1
idna==3.4
lavaplayer==1.0.10a0
multidict==6.0.4
mutagen==1.46.0
pycodestyle==2.10.0
pycryptodomex==3.16.0
python-dotenv==0.21.0
songbird-py==0.1.7
tomli==2.0.1
websockets==10.4
yarl==1.8.2
yt-dlp==2023.1.2
